requirements.txt
asgiref==3.3.4
autopep8==1.5.6
certifi==2020.12.5
Django==3.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
Pillow==8.2.0
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wincertstore==0.2
